[JIT] Make STORE polymorphic on input
We can always tile a STORE of a NUM operand correctly; in fact, the
tiler can always tile generic operands if have the same input and
output types.

The complexity of maintaining whether we need to store a num or a reg
value in the expression builder wasn't really worth it.