experiment to develop new heapsnapshot format
don't forget to add -lz to the cflags and ldflags.

this patch writes out all data in the heap snapshot
collection as separate files (one per "field") into
/tmp compressed through gzip.

there's a branch in the heapanalyzer repository that
turns these raw data files into the newly developed
format. The perl6 code in there will be the basis of
a C implementation inside moarvm.